Remove original parser; make ANTLR parser the only parser in the code
