--- Tests/TestCases/SaveTest.json
@@ -1,12 +1,0 @@
-[
-	{
-		"title": "Start",
-		"tags": "",
-		"body": "This is the first line.\r\nThis is the second line.\r\nThis is the third line.\r\nThis is the fourth line.\r\n",
-		"position": {
-			"x": 543,
-			"y": 269
-		},
-		"colorID": 0
-	}
-]
